% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{calcofi_map}
\alias{calcofi_map}
\title{calcofi_map}
\usage{
calcofi_map(
 
    geo = "https://raw.githubusercontent.com/marinebon/calcofi-analysis/master/data/plys_cinms.geojson",
  filter_str = "ply_code != \\"SoCal\\"",
  colors = c("red", "yellow")
)
}
\arguments{
\item{colors}{}
}
\value{

}
\description{
calcofi_map
}
