% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{calcofi_plot}
\alias{calcofi_plot}
\title{calcofi_plot}
\usage{
calcofi_plot(
  csv,
  x_fld = "year",
  y_fld = "avg_larvae_count_per_volume_sampled",
  y_trans = "log(y + 1)",
  x_lab = "Year",
  y_lab = "ln(mean abundance + 1)",
  title = NULL,
  yrs_recent = 5,
  interactive = T,
  in_loop = F
)
}
\arguments{
\item{in_loop}{}
}
\value{

}
\description{
calcofi_plot
}
