% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{calculate_statistics}
\alias{calculate_statistics}
\title{Generate statistics for any missing months for NMS Sanctuary}
\usage{
calculate_statistics(sanctuary, erddap_id, metric, csv_file)
}
\arguments{
\item{csv_file}{the csv file containing the data for the given metric for the sanctuary}

\item{wrapper_sanctuary}{code for national marine sanctuary}

\item{wrapper_erddap_id}{the name of the satellite data set to be pulled from erddap servers}

\item{wrapper_metric}{the parameter to be pulled from the data set}
}
\value{
nothing
}
\description{
Generate statistics for any missing months for NMS Sanctuary
}
