% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{generate_html_4_noninteractive_rmd}
\alias{generate_html_4_noninteractive_rmd}
\title{Generate the html for rmd files with non-interactive figures}
\usage{
generate_html_4_noninteractive_rmd(nms)
}
\arguments{
\item{nms}{The NMS sanctuary with only "cinms" currently doing anything.}
}
\value{
The function outputs a html file for every rmd file not containing interactive figures.
}
\description{
The purpose of this function is to insert the glossary tooltips into the html
for a rmd file (not containing interactive figures, which are dealt with by another
function as they present special complications). The function works by converting
a rmd file to a markdown file, inserting the relevant tooltip tags and scripts
into that markdown file, and then creating a html file from that markdown file.
}
\examples{
generate_html_4_noninteractive_rmd("cinms")

}
