% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{generate_latest_SST}
\alias{generate_latest_SST}
\title{Generate statistics for last month for NMS polygon}
\usage{
generate_latest_SST(
  wrapper_sanctuary_code,
  wrapper_erddap_id,
  wrapper_erddap_fld,
  wrapper_stats
)
}
\arguments{
\item{wrapper_sanctuary_code}{code for national marine sanctuary}

\item{wrapper_erddap_id}{the name of the satellite data set to be pulled from erddap servers}

\item{wrapper_erddap_fld}{the parameter to be pulled from the data set}

\item{wrapper_stats}{the statistics to be generated for the data set}
}
\value{
nothing
}
\description{
Generate statistics for last month for NMS polygon
}
