% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{get_figure_info}
\alias{get_figure_info}
\title{get_figure_info}
\usage{
get_figure_info(figure_id)
}
\arguments{
\item{figure_id}{}
}
\value{

}
\description{
The purpose of this function is to generate the hyperlinks for the monitoring program and data
associated with a figure and then to insert them into a gray bar above the figure in the modal window.
}
\examples{
calculate_statistics("cinms", "jplMURSST41mday", "sst", "avg-sst_cinms.csv")
calculate_statistics("cinms", "nesdisVHNSQchlaMonthly", "chlor_a", "avg-chl_cinms.csv")
}
