% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{get_modal_info}
\alias{get_modal_info}
\title{get_modal_info}
\usage{
get_modal_info(
  rmd = knitr::current_input(),
 
    info_modal_links_csv = "https://docs.google.com/spreadsheets/d/1yEuI7BT9fJEcGAFNPM0mCq16nFsbn0b-bNirYPU5W8c/gviz/tq?tqx=out:csv&sheet=info_modal_links"
)
}
\arguments{
\item{info_modal_links_csv}{}
}
\value{

}
\description{
get_modal_info
}
