% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{get_nms_polygon}
\alias{get_nms_polygon}
\title{Get NMS polygon}
\usage{
get_nms_polygon(nms)
}
\arguments{
\item{nms}{code for national marine sanctuary}

\item{dir_shp}{directory to store cached shapefile}
}
\value{
sf object
}
\description{
given NMS code (see
\url{https://sanctuaries.noaa.gov/library/imast_gis.html}), download and
extract zip, cache shapefile or read existing shapefile
}
