% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{get_nms_polygons}
\alias{get_nms_polygons}
\title{Get NMS polygons}
\usage{
get_nms_polygons(nms)
}
\arguments{
\item{nms}{code for national marine sanctuary}
}
\value{
sf object
}
\description{
given NMS code (see
\url{https://sanctuaries.noaa.gov/library/imast_gis.html}), download and
extract zip, cache shapefile or read existing shapefile
}
