% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{glossarize_md}
\alias{glossarize_md}
\title{Insert html tags for glossary tooltips into md files}
\usage{
glossarize_md(md, md_out = md)
}
\arguments{
\item{md}{the md file where the tags are to be inserted}

\item{md_out}{the md output file}
}
\value{
nothing
}
\description{
The purpose of this function is to insert the html tags required for glossary tooltip functionality into
a given md file
}
