% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{initial_build}
\alias{initial_build}
\title{Create new site from template}
\usage{
initial_build(directory_name)
}
\arguments{
\item{directory_name}{The directory where the new site should be located.}
}
\value{
The output is a directory containing the template version of a site.
}
\description{
This function creates a new site from a template stored within the package.
}
\examples{
\dontrun{
directory_name("test_location")
}

}
