% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_data.R
\name{make_rocky_sites}
\alias{make_rocky_sites}
\title{Make Sanctuary sites for mapping from raw rocky intertidal (MARINe) data file}
\usage{
make_rocky_sites(
  raw_csv,
  nms_ply = NULL,
  site_regions = NULL,
  raw_source = c("csv", "erddap")
)
}
\arguments{
\item{raw_csv}{raw data file containing fields: \code{marine_site_name}, \code{longitude}, \code{latitude}.}

\item{nms_ply}{optional National Marine Sanctuary polygon, e.g. from \code{get_nms_polygons("ocnms")} to spatially filter from \code{raw_csv} within a 0.01 decimal degrees (≈ 1.11 km). Usually include either parameter \code{nms_ply} or \code{site_regions} to select sites for a given sanctuary.}

\item{site_regions}{optional lookup table containing fields for region \code{rgn} and site \code{site} to filter by \code{site} from \code{raw_csv} and append \code{rgn} field. Usually include either parameter \code{nms_ply} or \code{site_regions} to select sites for a given sanctuary.}

\item{raw_source}{either "csv" (default) or "erddap" for handling file format differences}
}
\value{
sites A data frame with 59 rows and 4 variables:
\describe{
\item{nms}{National Marine Sanctuary code, e.g. CINMS or MBNMS}
\item{rgn}{region}
\item{site}{site}
}
}
\description{
Make Sanctuary sites for mapping from raw rocky intertidal (MARINe) data file
}
