% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{md_caption}
\alias{md_caption}
\title{md_caption}
\usage{
md_caption(title, md = here::here("modals/_captions.md"), get_details = F)
}
\arguments{
\item{get_details}{}
}
\value{

}
\description{
md_caption
}
