% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{plot_SST_anomaly}
\alias{plot_SST_anomaly}
\title{Plot SST anomaly data}
\usage{
plot_SST_anomaly(csv_SST)
}
\arguments{
\item{csv_SST}{The csv file containing the SST anomaly data to be plotted.}
}
\value{
This function outputs a dygraph object of the time series plot.
}
\description{
The purpose of this function is to plot the sea surface temperature anomaly
data.
}
\examples{
\dontrun{
csv_SST <-here::here("data/oceano/sst_anomaly_cinms.csv")
plot_metric_timeseries(csv_SST)
}

}
