% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{plot_timeseries}
\usage{
plot_timeseries(d, title = "SST", color = "red", dyRangeSelector = T, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
plot_timeseries
}
