% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{ply2erddap}
\alias{ply2erddap}
\title{Generate statistics for raster data within NMS polygon}
\usage{
ply2erddap(sanctuary_code, erddap_id, erddap_fld, year, month, stats)
}
\arguments{
\item{sanctuary_code}{code for national marine sanctuary}

\item{erddap_id}{the name of the satellite data set to be pulled from erddap servers}

\item{erddap_fld}{the parameter to be pulled from the data set}

\item{year}{year of requested data in integer form}

\item{month}{month of requested data in integer form}

\item{stats}{the statistics to be generated for the data set}
}
\value{
statistics
}
\description{
Generate statistics for raster data within NMS polygon
}
