% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{read_csv_fmt}
\alias{read_csv_fmt}
\title{read_csv_fmt BEN}
\usage{
read_csv_fmt(csv, erddap_format = "csv")
}
\arguments{
\item{csv}{BEN}

\item{erddap_format}{BEN}
}
\value{
BEN
}
\description{
description BEN
}
