% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{render_all_rmd}
\alias{render_all_rmd}
\title{Render all rmd files in the modals folder}
\usage{
render_all_rmd(nms = "cinms", interactive_only = F)
}
\arguments{
\item{nms}{The NMS sanctuary with only "cinms" currently doing anything.}

\item{interactive_only}{A Boolean variable indicating whether only rmd files containing interactive figures should be rendered.}
}
\value{
The function outputs a html file for every rmd file in the modals folder.
}
\description{
Render all rmd files in the modals folder
}
\examples{
render_all_rmd(interactive_only = T)

}
