% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{render_modal_windows}
\alias{render_modal_windows}
\title{Render html for all R Markdown files in modal directory}
\usage{
render_modal_windows(NMS)
}
\arguments{
\item{NMS}{the NMS sanctuary}
}
\value{
This function outputs html files for most rmd files in the modal directory.
}
\description{
This function generates the full html, including tooltips, for all R markdown
files in the modal directory.
}
\examples{
render_modal_windows("cinms")

}
