% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{rmd2html}
\alias{rmd2html}
\title{Render html for rmd files, including glossary tooltips}
\usage{
rmd2html(rmd)
}
\arguments{
\item{rmd}{the rmd file to be rendered into html}
}
\value{
nothing
}
\description{
Render html for rmd files, including glossary tooltips
}
