% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rocky_clusters}
\alias{rocky_clusters}
\title{Rocky intertidal site Sanctuary groupings}
\format{
A tibble with the following columns:
\describe{
\item{nms}{National Marine Sanctuary acronym, e.g. CINMS}
\item{cluster}{group to name cluster of sites, e.g. Santa Cruz Island}
\item{site}{site in rocky_sanctuary_data}
}
}
\source{
\url{https://marine.ucsc.edu/sites/sites-region/index.html}
}
\usage{
rocky_clusters
}
\description{
A simple table describing groupings per National Marine Sanctuary.
}
\keyword{datasets}
