% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rocky_counts}
\alias{rocky_counts}
\title{Rocky intertidal species counts}
\format{
A simple features tibble with:
\describe{
\item{nms}{National Marine Sanctuary acronym, e.g. CINMS}
\item{cluster}{grouping of sites, e.g. Santa Cruz Island}
\item{site}{MARINe site}
\item{date}{annual date YYYY-06-15}
\item{sp_method}{MARINe method used for sampling the plot:
IP = Irregular Plot; BT25 = Band Transect 2m x 5m;
GSES = General search entire site; TS30 = Timed Search 30 Minute.}
\item{sp_code}{MARINe species code}
\item{count}{average count}
}
}
\source{
\url{https://marine.ucsc.edu/sites/sites-region/index.html}
}
\usage{
rocky_counts
}
\description{
A table of MARINe intertidal average species counts summarized annually
across sanctuaries, clusters (including "ALL"), sites (including "ALL"),
sp_method (including "ALL") and species code \code{sp_code}. All data has sp_target (or
target_assemblage) of "sea_star".
}
\keyword{datasets}
