% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rocky_cover}
\alias{rocky_cover}
\title{Rocky intertidal percent cover data}
\format{
A simple features tibble with:
\describe{
\item{nms}{National Marine Sanctuary acronym, e.g. CINMS}
\item{cluster}{grouping of sites, e.g. Santa Cruz Island}
\item{site}{MARINe site}
\item{date}{annual date YYYY-06-15}
\item{target_assemblage}{MARINe target assemblage}
\item{sp_code}{MARINe species code}
\item{pct_cover}{percent cover}
}
}
\source{
\url{https://marine.ucsc.edu/sites/sites-region/index.html}
}
\usage{
rocky_cover
}
\description{
A big table of MARINe intertidal percent cover data summarized annually
across sanctuaries, clusters (including "ALL"), sites (including "ALL"),
target_assemblage (including "ALL") and species code.
}
\keyword{datasets}
