% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocky.R
\name{rocky_map}
\alias{rocky_map}
\title{Map rocky intertidal sites where data was collected}
\usage{
rocky_map(nms)
}
\arguments{
\item{nms}{The National Marine Sanctuary code.}
}
\value{
This function returns a mapview object displaying data collection sites.
}
\description{
This function produces an interactive map showing where rocky intertidal data
was collected by the MARINe consortium.
}
\examples{
\dontrun{
rocky_map("cinms")
}

}
