% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rocky_sites}
\alias{rocky_sites}
\title{Rocky intertidal sites}
\format{
A simple features tibble with:
\describe{
\item{site}{site}
\item{lat}{latitude}
\item{lon}{longitude}
\item{nms}{National Marine Sanctuary acronym, e.g. CINMS}
\item{cluster}{group to name cluster of sites, e.g. Santa Cruz Island}
}
}
\source{
\url{https://marine.ucsc.edu/sites/sites-region/index.html}
}
\usage{
rocky_sites
}
\description{
A simple features table of MARINe intertidal sites associated with a National Marine Sanctuary.
}
\keyword{datasets}
