% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rocky_species}
\alias{rocky_species}
\title{Rocky species}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 70 rows and 12 columns.
}
\source{
\url{https://marine.ucsc.edu/sites/sites-region/index.html}
}
\usage{
rocky_species
}
\description{
A table of MARINe species with \code{sp_code} to relate to \link{rocky_cover}
and \link{rocky_counts}.
}
\keyword{datasets}
