% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{calcofi_plot}
\alias{calcofi_plot}
\title{Produces plots of CALCOFI data.}
\usage{
calcofi_plot(
  csv,
  x_fld = "year",
  y_fld = "avg_larvae_count_per_volume_sampled",
  y_trans = "log(y + 1)",
  x_lab = "Year",
  y_lab = "ln(mean abundance + 1)",
  title = NULL,
  yrs_recent = 5,
  interactive = T,
  in_loop = F
)
}
\arguments{
\item{csv}{The CALCOFI time series data, in csv format, to be plotted.}

\item{x_fld}{The column of the time series data to be used for the x-axis of the plot.}

\item{y_fld}{The column of the time series data to be used for the y-axis of the plot.}

\item{y_trans}{The transformation to be performed on the data used for the y-axis.}

\item{x_lab}{The label for the x-axis on the plot.}

\item{y_lab}{The label for the y-axis on the plot.}

\item{title}{The label for the title of the plot.}

\item{yrs_recent}{The number of most recent years to be shaded in the plot.}

\item{interactive}{A Boolean variable indicating whether the plot is to be interactive or not.}

\item{in_loop}{A Boolean variable indicating whether an error condition exists.}
}
\value{
The output is a plot of time series data.
}
\description{
This function produces plots of CALCOFI-originated time series data.
}
\examples{
calcofi_plot(csv = "https://raw.githubusercontent.com/marinebon/calcofi-analysis/master/data/Anchovy_CINMS.csv", title = "Anchovy - CINMS Region")

}
