% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_data.R
\name{calculate_SST_anomaly}
\alias{calculate_SST_anomaly}
\title{Calculate SST anomaly}
\usage{
calculate_SST_anomaly(sanct)
}
\arguments{
\item{sanct}{The NMS sanctuary, with only the value "cinms" currently doing anything.}
}
\value{
The output is a csv file containing the time series anomaly data.
}
\description{
This function calculates the SST anomaly for every month of SST data and then
writes that out to a csv file that will later be used to produce a SST anomaly figure. The
approach used here is based on \href{https://www.nature.com/articles/ncomms13757}{Reed et al. 2016, Nature Communications}.
Using the first 15 full years of data from the SST data set (2003-2017), an average SST value is
generated for every month of the year. Then the average SST value for the appropriate month is
subtracted from every value in the SST dataset.
}
\examples{
\dontrun{
calculate_SST_anomaly("cinms")
}

}
