% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{generate_html_4_noninteractive_rmd}
\alias{generate_html_4_noninteractive_rmd}
\title{This function generates the html for rmd files with non-interactive figures.}
\usage{
generate_html_4_noninteractive_rmd(nms)
}
\arguments{
\item{nms}{the NMS sanctuary}
}
\value{
nothing
}
\description{
This function generates the html for rmd files with non-interactive figures.
}
