% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{get_nms_polygons}
\alias{get_nms_polygons}
\title{Get National Marine Sanctuary polygons}
\usage{
get_nms_polygons(nms)
}
\arguments{
\item{nms}{The code for a national marine sanctuary.}
}
\value{
The function returns a sf object containing the polygons of a sanctuary.
}
\description{
Given NMS code (see
\url{https://sanctuaries.noaa.gov/library/imast_gis.html}), download and
extract zip, cache shapefile or read existing shapefile.
}
\examples{
get_nms_polygons("cinms")

}
