% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{get_nms_polygons}
\alias{get_nms_polygons}
\title{Get NMS polygon}
\usage{
get_nms_polygons(nms)
}
\arguments{
\item{nms}{code for national marine sanctuary}

\item{dir_shp}{directory to store cached shapefile}

\item{sanctuary_code}{code for national marine sanctuary}

\item{erddap_id}{the name of the satellite data set to be pulled from erddap servers}

\item{erddap_fld}{the parameter to be pulled from the data set}

\item{year}{year of requested data in integer form}

\item{month}{month of requested data in integer form}

\item{stats}{the statistics to be generated for the data set}
}
\value{
sf object

statistics
}
\description{
given NMS code (see
\url{https://sanctuaries.noaa.gov/library/imast_gis.html}), download and
extract zip, cache shapefile or read existing shapefile
}
\examples{
Generate statistics for raster data within NMS polygon

}
