% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{plot_intertidal_nms}
\alias{plot_intertidal_nms}
\title{plot_intertidal_nms}
\usage{
plot_intertidal_nms(
  d_csv,
  NMS,
  spp,
  sp_name,
  spp_targets = NULL,
  fld_val = "pct_cover",
  label_y = "Annual Mean Percent Cover (\%)",
  label_x = "Year",
  nms_skip_regions = c("OCNMS", "MBNMS")
)
}
\arguments{
\item{nms_skip_regions}{}
}
\value{

}
\description{
plot_intertidal_nms
}
