% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{plot_metric_timeseries}
\alias{plot_metric_timeseries}
\title{plot_metric_timeseries}
\usage{
plot_metric_timeseries(csv, metric, ...)
}
\arguments{
\item{...}{additional parameters to pass to \link[dygraphs]{dygraph}}
}
\value{

}
\description{
plot_metric_timeseries
}
