% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{ply2erddap}
\alias{ply2erddap}
\title{Extract ERDDAP data from polygon}
\usage{
ply2erddap(sanctuary_code, erddap_id, erddap_fld, year, month, stats)
}
\arguments{
\item{stats}{}
}
\value{

}
\description{
Extract ERDDAP data from polygon
}
