% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{render_modal_windows}
\alias{render_modal_windows}
\title{This function generates the html for all rmd files in the modal directory.}
\usage{
render_modal_windows(NMS)
}
\arguments{
\item{NMS}{the NMS sanctuary}
}
\value{
nothing
}
\description{
This function generates the html for all rmd files in the modal directory.
}
